+++ server/src/main/kotlin/com/coolcode/server/dto/evaluation/EvaluationResult.kt
@@ -1,9 +1,7 @@
 package com.coolcode.server.dto.evaluation
 
-import java.math.BigDecimal
-
 data class EvaluationResult(
     val runId: String,
-    val score: BigDecimal,
+    val score: Int,
     val message: String
 )
